Consolidate FixedVector into a single module
